refactoring identificators (Int -> Long)

+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/data/Mappers.kt
@@ -6,7 +6,7 @@ import com.spinoza.messenger_tfs.domain.model.Message
 import com.spinoza.messenger_tfs.domain.model.ReactionParam
 import java.util.*
 
-fun MessageDto.toDomain(userId: Int): Message {
+fun MessageDto.toDomain(userId: Long): Message {
     return Message(
         date = this.date,
         userId = this.userId,
@@ -19,15 +19,15 @@ fun MessageDto.toDomain(userId: Int): Message {
     )
 }
 
-fun TreeSet<MessageDto>.toDomain(userId: Int): List<Message> {
+fun TreeSet<MessageDto>.toDomain(userId: Long): List<Message> {
     return this.map { it.toDomain(userId) }
 }
 
-private fun ReactionParamDto.toDomain(userId: Int): ReactionParam {
+private fun ReactionParamDto.toDomain(userId: Long): ReactionParam {
     return ReactionParam(this.usersIds.size, this.usersIds.contains(userId))
 }
 
-fun Message.toDto(userId: Int, messageId: Int): MessageDto {
+fun Message.toDto(userId: Long, messageId: Long): MessageDto {
     return MessageDto(
         date = this.date,
         userId = this.userId,
@@ -39,10 +39,10 @@ fun Message.toDto(userId: Int, messageId: Int): MessageDto {
     )
 }
 
-private fun Map<String, ReactionParamDto>.toDomain(userId: Int): Map<String, ReactionParam> {
+private fun Map<String, ReactionParamDto>.toDomain(userId: Long): Map<String, ReactionParam> {
     return this.map { it.key to it.value.toDomain(userId) }.toMap()
 }
 
-private fun Map<String, ReactionParam>.toDto(userId: Int): Map<String, ReactionParamDto> {
+private fun Map<String, ReactionParam>.toDto(userId: Long): Map<String, ReactionParamDto> {
     return this.map { it.key to ReactionParamDto(listOf(userId)) }.toMap()
 }
