+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/presentation/adapter/delegate/DelegateItem.kt
@@ -4,7 +4,7 @@ interface DelegateItem {
 
     fun content(): Any
 
-    fun id(): Int
+    fun id(): Long
 
     fun compareToOther(other: DelegateItem): Boolean
 }
